// packages/workbench-core/src/core/types.ts
export interface SchedulerInfo {
  key: string;
  name: string;
  queueName: string;
  pattern?: string;
  every?: number;
  next?: number;
  endDate?: number;
  tz?: string;
}